// frontend/src/shared/types/theme.ts
export type Theme =
    | 'ayu-dark'
    | 'ayu-light'
    | 'ayu-mirage'
    | 'bear-theme'
    | 'bluloco-dark'
    | 'bluloco-light'
    | 'catppuccin'
    | 'catppuccin-frappe'
    | 'catppuccin-latte'
    | 'catppuccin-macchiato'
    | 'cobalt2'
    | 'cyberpunk'
    | 'cyberpunk-light'
    | 'dark'
    | 'dracula'
    | 'dracula-light'
    | 'everforest'
    | 'everforest-light'
    | 'firefly-pro'
    | 'github-dark'
    | 'github-light'
    | 'gruvbox-dark'
    | 'gruvbox-light'
    | 'horizon-dark'
    | 'horizon-light'
    | 'light'
    | 'lucario'
    | 'material-darker'
    | 'material-ocean'
    | 'minimal-dark'
    | 'minimal-light'
    | 'monokai'
    | 'monokai-light'
    | 'monokai-pro'
    | 'night-owl'
    | 'night-owl-light'
    | 'noctis'
    | 'nord'
    | 'nord-light'
    | 'one-dark'
    | 'one-light'
    | 'one-monokai'
    | 'palenight'
    | 'palenight-light'
    | 'panda'
    | 'poimandres'
    | 'quiet-light'
    | 'rose-pine'
    | 'rose-pine-dawn'
    | 'rose-pine-moon'
    | 'shades-of-purple'
    | 'solarized-dark'
    | 'solarized-light'
    | 'sublime-material'
    | 'synthwave'
    | 'synthwave-light'
    | 'tokyo-night'
    | 'tokyo-night-light'
    | 'winter-is-coming-dark'
    | 'winter-is-coming-light'